monitoring/tracer/requirements.txt
pyyaml==5.3.1
flask==1.1.2
gunicorn==20.0.4
-r ../monitorlib/requirements.txt
